// CLIENT/src/app/models/feedbackResponseBean.ts
// feedback-response-bean.model.ts
export class FeedbackResponseBean {
  id: number | undefined;
  suggestionType: string | undefined;
  adminId: number | undefined;
  feedbackReply: string | undefined;
  feedbackText: string | undefined;
  userId: number | undefined;
  rating: number | undefined;
}